requirements.txt
Flask
flask-cors
requests
psycopg2-binary
gunicorn
